-Fixes the throne issue for the king -Raises the value of gold objects slightly, for consistency -Gives a few of the classes (particularly travelling warriors) golden burial rings (or silver for the unlucky). -comments out dildos from former codebase (teehee oversight)
Essentially building up the skeleton for burial rings. In old Christian tradition, sailors would wear a golden ring to pay for a Christian burial. I can't code earrings yet, so rings will do. The plan is to later find a way to curse the ring and make it worth 0 until the original holder is buried, but for now we leave it up to players who find the body (which actually plays into roleplay, hm).